refactor(physicalmem): unify physical memory detection
