+++ src/arch/aarch64/mm/mod.rs
@@ -1,12 +1,11 @@
 pub mod paging;
-pub mod physicalmem;
-
-pub use self::physicalmem::init_page_tables;
 
 pub fn init() {
 	unsafe {
 		paging::init();
 	}
-	physicalmem::init();
+	crate::mm::physicalmem::init();
 	crate::mm::virtualmem::init();
 }
+
+pub fn init_page_tables() {}
