+++ src/arch/riscv64/mm/mod.rs
@@ -1,9 +1,8 @@
 pub mod paging;
-pub mod physicalmem;
 
 pub use self::paging::init_page_tables;
 
 pub fn init() {
-	physicalmem::init();
+	crate::mm::physicalmem::init();
 	crate::mm::virtualmem::init();
 }
