+++ src/arch/x86_64/mm/mod.rs
@@ -1,5 +1,4 @@
 pub(crate) mod paging;
-pub(crate) mod physicalmem;
 
 use memory_addresses::arch::x86_64::{PhysAddr, VirtAddr};
 #[cfg(feature = "common-os")]
@@ -59,7 +58,7 @@ pub fn create_new_root_page_table() -> usize {
 
 pub fn init() {
 	paging::init();
-	physicalmem::init();
+	crate::mm::physicalmem::init();
 	crate::mm::virtualmem::init();
 
 	#[cfg(feature = "common-os")]
